+++ packages/client/CHANGELOG.md
@@ -31,8 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - resends support multiple storage nodes (the data is fetched from a random storage node)
 - Configuration parameter `groupKeys` renamed to `encryptionKeys`
 - Exported classes `GroupKey` and `GroupKeyId` renamed to `EncryptionKey` and `EncryptionKeyId`
-- Configuration parameter `verifySignatures` changed from enum to boolean
-  - to disable signature verification, use `verifySignatures=false` (previously `verifySignatures=never`)
 
 ### Deprecated
 
@@ -45,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove method `unsubscribeAll()`, use `unsubscribe()` without arguments instead
 - Remove client configuration option `client.network.name`
 - Remove `subscription.onMessage`, `onStart` and `onError` methods, use `subscription.on('error', cb)` to add an error listener
+- Remove client configuration option `verifySignatures`
 
 ### Fixed